net: qrtr: ns: Change servers radix tree to xarray
There is a use after free scenario while iterating through the servers
radix tree despite the ns being a single threaded process. This can
happen when the radix tree APIs are not synchronized with the
rcu_read_lock() APIs.

Convert the radix tree for servers to xarray to take advantage of the
built in rcu lock usage provided by xarray.

Change-Id: Ib4f11bb432258f4050c633ad73b4e8110fcc2758
CRs-Fixed: 3716307
Signed-off-by: Chris Lew <clew@codeaurora.org>
Signed-off-by: Jay Jayanna <jayanna@codeaurora.org>
Signed-off-by: Jainuddin Shaik <quic_jshaik@quicinc.com>